Refactor the code-base to be a poly development project
I want to try and convert this whole thing into a poly project, as a
means of refactoring it and as a way to better learn poly.

Also, I think there are advantages to doing it this way, so that I can
cleanly demarcate puclic data generated from Org-mode vs private data
vs visual projects.